Make all_reasons switchable.
I initially had this code switched off by default, and
only had it forced on so I could test it, and then forgot to switch it
off again. However, it has proved very useful to have it on by default.

So instead, I've left it as on by default, but eliminated the sillyness
where it was impossible to turn off.